feat: implement errors Handling

+++ Application/Activities/Commands/DeleteActivities.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Application.Core;
 using Domain;
 using MediatR;
 using Persistence;
@@ -10,22 +11,27 @@ namespace Application.Activities.Commands
 {
   public class DeleteActivities
   {
-    public class Command : IRequest<Guid>
+    public class Command : IRequest<Result<Unit>>
     {
       public required Guid Id { get; set; }
     }
 
-    public class Handler(AppDbContext context) : IRequestHandler<Command, Guid>
+    public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>>
     {
 
-      public async Task<Guid> Handle(Command request, CancellationToken cancellationToken)
+      public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
       {
-        var activity = await context.Activities.FindAsync(request.Id)?? throw new Exception("Activity not found");
+        var activity = await context.Activities.FindAsync(request.Id);
+
+        if (activity == null) return Result<Unit>.Failure("Activity not found", 404);
 
         context.Activities.Remove(activity);
-        await context.SaveChangesAsync(cancellationToken);
 
-        return activity.Id;
+       var result = await context.SaveChangesAsync(cancellationToken);
+
+       if (result > 0) return Result<Unit>.Success(Unit.Value);
+
+       return Result<Unit>.Failure("Failed to delete activity", 400);
       }
   } }
 }
